Relocated `trace_locality` to `RefBundle`
