+++ python/ray/data/tests/test_ref_bundle.py
@@ -1,5 +1,6 @@
 from unittest.mock import patch
 
+import ray
 from ray import ObjectRef
 from ray.data._internal.execution.interfaces import RefBundle
 from ray.data.block import BlockMetadata
@@ -52,3 +53,37 @@ def test_get_preferred_locations():
         "2": 7168,  # first_block_ref, second_block_ref, third_block_ref
         "3": 3072,  # first_block_ref, second_block_ref
     } == preferred_object_locs
+
+
+def test_calculate_ref_hits(ray_start_regular_shared):
+    first_block_ref = ObjectRef(b"1" * 28)
+    second_block_ref = ObjectRef(b"2" * 28)
+    third_block_ref = ObjectRef(b"3" * 28)
+
+    meta = BlockMetadata(num_rows=None, size_bytes=1, exec_stats=None, input_files=None)
+    bundle = RefBundle(
+        blocks=[
+            (first_block_ref, meta),
+            (second_block_ref, meta),
+            (third_block_ref, meta),
+        ],
+        owns_blocks=True,
+        schema=None,
+    )
+
+    # NOTE: Setting and retrieving this setting is kept purely for
+    #       compatibility purposes. This config has no impact otherwise
+    ctx = ray.data.context.DataContext.get_current()
+    prev_enable_get_object_locations_for_metrics = (
+        ctx.enable_get_object_locations_for_metrics
+    )
+    try:
+        ctx.enable_get_object_locations_for_metrics = True
+        hits, misses, unknowns = bundle.trace_locality(ray.get_runtime_context().get_node_id())
+        assert hits == 2
+        assert misses == 0
+        assert unknowns == 0
+    finally:
+        ctx.enable_get_object_locations_for_metrics = (
+            prev_enable_get_object_locations_for_metrics
+        )
